Refactor out common slack job utilities
